transferpy: Generate checksum parallel to the data transfer
In the current scenario, the checksum can be enabled only for the type
`file`, and the time taken for the data transfer is approximately equal to
(the number of destination hosts + 1)*(one checksum calculation time)+
(the actual transfer time).

This patch gives users a new option `--parallel-checksum,` which makes
the process faster by allowing the transferpy to calculate the checksum
parallel to the actual data transfer. This option can be enabled for all
three transfer types `file`, `xtrabackup`, and `decompress`. Since the
checksum calculation happens parallel to the transfer, the total time taken
to complete the data transfer will be similar to the case without checksum.
Since the checksum calculation happens just before the actual data transfer,
this can check for the network-related integrity issues, not the file
system related issues.

This patch adds the `tee >(md5sum > checksumPath)` command to the
src/dst_command in the copy_to function.

Bug: T254979
Change-Id: I8c75570162c972f188a6474dd396e309c35be5f3